Librarian refactoring

+++ src/com/library/humans/Librarian.kt
@@ -16,12 +16,26 @@ class Librarian(
     override val startMessage: String = "`${this.name} ${this.surname}` на рабочем месте"
     override val endMessage: String = "`${this.name} ${this.surname}` окончила работу"
 
+    override fun onResult(result: Result) {
+        when (result) {
+            is Result.BookIsConfiscated -> {
+                if (repository.wishBooks(result.client).isNotEmpty()) {
+                    giveBooksIfPossible(result.client)
+                } else {
+                    offerBook(result.client)
+                }
+            }
+            Result.BookOffered -> Unit
+            is Result.ReleaseTheClient -> releaseTheClient(result.client)
+        }
+    }
+
     fun confiscateBooks(client: Client) {
         this.workWithClient.set(true)
         addTask(Task.ConfiscateBooks(client = client, librarian = this, repository = repository))
     }
 
-    fun giveBookIfPossible(client: Client) {
+    fun giveBooksIfPossible(client: Client) {
         this.workWithClient.set(true)
         addTask(Task.GiveABookIfYouHaveOne(
             client = client,
@@ -36,32 +50,12 @@ class Librarian(
     }
 
     fun giveRandomBook(client: Client) {
-        addTask(Task.GiveRandomBook(client = client, librarian = this, repository = repository))
+        addTask(Task.GiveRandomBook(client = client, repository = repository))
     }
 
-    fun releaseTheClient(client: Client) {
+    private fun releaseTheClient(client: Client) {
         client.finish()
         workWithClient.set(false)
-    }
-
-    override fun onResult(result: Result) {
-        when (result) {
-            is Result.BookInTheLibrary -> {
-                if (repository.wishBooks(result.client).isNotEmpty()) {
-                    giveBookIfPossible(result.client)
-                } else {
-                    offerBook(result.client)
-                }
-            }
-            Result.OfferABook -> Unit
-            is Result.BookHanded -> {
-                if (repository.wishBooks(result.client).isNotEmpty()) {
-                    offerBook(result.client)
-                } else {
-                    releaseTheClient(result.client)
-                }
-            }
-        }
     }
 
     sealed class Task : Human.Task<Result> {
@@ -77,7 +71,7 @@ class Librarian(
                     "Библиотекарь `${librarian.name} ${librarian.surname}` " +
                             "забрала книги у `${client.name} ${client.surname}`"
                 )
-                return Result.BookInTheLibrary(client)
+                return Result.BookIsConfiscated(client)
             }
         }
 
@@ -87,20 +81,27 @@ class Librarian(
             private val repository: Repository,
         ) : Task() {
             override fun execute(): Result {
-                repository.wishBooks(client).forEach { necessaryBook: Book ->
-                    println("`${client.name} ${client.surname}` хочет взять книгу `$necessaryBook`.")
-                    val isAvailable = repository.isBookAvailable(necessaryBook)
-                    if (isAvailable) {
-                        repository.addBookToClient(client, necessaryBook)
-                        repository.removeFromWishes(client, necessaryBook)
-                        println(
-                            "Библиотекарь `${librarian.name} ${librarian.surname}`" +
-                                    "посетителю `${client.name} ${client.surname}`"
-                        )
-                    } else println("Книга `$necessaryBook` отсутствует.")
+                val wishBooks = repository.wishBooks(client)
+                if (wishBooks.isEmpty()) {
+                    println("Список желаемых книг пуст.")
+                    return Result.ReleaseTheClient(client)
+                } else {
+                    wishBooks.forEach { necessaryBook: Book ->
+                        println("`${client.name} ${client.surname}` хочет взять книгу `$necessaryBook`.")
+                        val isAvailable = repository.isBookAvailable(necessaryBook)
+                        if (isAvailable) {
+                            repository.addBookToClient(client, necessaryBook)
+                            repository.removeFromWishes(client, necessaryBook)
+                            println(
+                                "Библиотекарь `${librarian.name} ${librarian.surname}`" +
+                                        "посетителю `${client.name} ${client.surname}`"
+                            )
+                        } else {
+                            println("Книга `$necessaryBook` отсутствует.")
+                        }
+                    }
+                    return Result.ReleaseTheClient(client)
                 }
-
-                return Result.BookHanded(client)
             }
         }
 
@@ -110,34 +111,35 @@ class Librarian(
         ) : Task() {
             override fun execute(): Result {
                 println("Библиотекарь `${librarian.name} ${librarian.surname}` предлагает `${client.name} ${client.surname}` взять другую книгу")
-                client.decisionToBorrowABook(librarian)
-                return Result.OfferABook
+                val clientAgreed = client.decisionToBorrowABook(librarian)
+                return if (clientAgreed) {
+                    Result.BookOffered
+                } else {
+                    Result.ReleaseTheClient(client)
+                }
             }
         }
 
         class GiveRandomBook(
             private val client: Client,
-            private val librarian: Librarian,
             private val repository: Repository,
         ) : Task() {
             override fun execute(): Result {
                 val randomBook: Book? = repository.randomAvailableBookOrNull()
-
                 if (randomBook == null) {
                     println("К сожалению книг нет.")
-                    librarian.releaseTheClient(client)
                 } else {
                     repository.addBookToClient(client, randomBook)
                     println("`${client.name} ${client.surname}` взял предложенную книгу `${randomBook.title}`")
                 }
-                return Result.BookHanded(client)
+                return Result.ReleaseTheClient(client)
             }
         }
     }
 
     sealed class Result : Human.Task.Result {
-        class BookInTheLibrary(val client: Client) : Result()
-        class BookHanded(val client: Client) : Result()
-        object OfferABook : Result()
+        class BookIsConfiscated(val client: Client) : Result()
+        class ReleaseTheClient(val client: Client) : Result()
+        object BookOffered : Result()
     }
 }
